+++ apps/mobile/app.config.js
@@ -31,6 +31,6 @@ module.exports = {
     web: {
       favicon: './assets/favicon.png',
     },
-    plugins: ['expo-router'],
+    plugins: ['expo-router', 'expo-localization'],
   },
 };